feat: Add support for reading comment text from files
Added a new function `resolveCommentText` that checks if a comment body starts
with "@". If it does, it interprets the rest of the string as a file path and
reads the comment text from that file. This allows users to provide longer or
more complex comment bodies without having to type them directly on the command
line.

This change was implemented to improve the usability of the `resolve` command,
especially for cases where detailed explanations or code snippets are required
for resolving comments.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>